requirements.txt
black
click
jupyterlab
kafka-python
mypy-extensions
pathspec
platformdirs
psycopg2
tomli
typing-extensions
